+++ internal/app/kconfig-controller/controller/contant.go
@@ -3,7 +3,9 @@ package controller
 const (
 	controllerAgentName = "kconfig-controller"
 	// KconfigEnabledDeploymentAnnotation annotation for enabling deployment management (true/false)
-	KconfigEnabledDeploymentAnnotation = "kconfigcontroller.atteg.com/enabled"
+	KconfigEnabledDeploymentAnnotation = "kconfigcontroller.atteg.com/env"
+	// KconfigEnvRefVersionAnnotation deployment template annotation key where value equals current EnvRefVersion string to track update
+	KconfigEnvRefVersionAnnotation = "kconfigcontroller.atteg.com/envrefversions"
 	// SuccessSynced is used as part of the Event 'reason' when a Foo is synced
 	SuccessSynced = "Synced"
 	// ErrResourceExists is used as part of the Event 'reason' when a Foo fails